assignation entreprise et budget

+++ clouds5spring/src/main/java/com/carte/clouds5spring/repository/RouteEntrepriseRepository.java
@@ -3,5 +3,5 @@ package com.carte.clouds5spring.repository;
 import com.carte.clouds5spring.entity.RouteEntreprise;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RouteEntrepriseRepository extends JpaRepository<RouteEntreprise, Long> {
+public interface RouteEntrepriseRepository extends JpaRepository<RouteEntreprise, Integer> {
 }
